requirements.txt
# 字幕核对工具 - Python 依赖包
# 安装方法: pip install -r requirements.txt

# Eel - 用于创建桌面应用程序的 Python 库
# 它允许使用 HTML/CSS/JavaScript 作为前端，Python 作为后端
eel>=0.16.0

# python-docx - 用于读取和处理 Word (.docx) 文件
# 注意：仅支持 .docx 格式，不支持旧的 .doc 格式
python-docx>=1.1.0
